Move `gen_san_strings_new.py` -> `__init__.py`
